requirements.txt
diffusers==0.31.0
google-generativeai==0.8.3
matplotlib==3.9.3
pandas==2.2.3
plotly==5.24.1
pyarrow==18.1.0
python-dotenv==1.0.1
requests==2.32.3
streamlit==1.40.2
transformers==4.47.1
